ai-agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Agentic travel concierge with deals and concierge agents (FastAPI + Kafka + SQLModel)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.24",
  "pydantic>=2.5",
  "websockets>=12.0",
  "python-dotenv>=1.0",
  "httpx>=0.24",
  "pandas>=2.1",
  "numpy>=1.26",
  "orjson>=3.9",
  "langchain>=0.2.0",
  "langgraph>=0.0.46",
  "langchain-google-genai>=1.0.2",
  "pytest>=8.2",
  "pymongo>=4.8",
  "httpx>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.uv.sources]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "pyproject.toml", "requirements.txt", "README.md"]
